requirements.txt
django ~= 4.2
Faker
django-crispy-forms
crispy-bootstrap5
django-tinymce
Pillow
